# re-write flutter analyze (the single-shot and --flutter-repo) to use the analysis server