--- a/packages/flutter_tools/test/commands/analyze_once_test.dart
+++ b/packages/flutter_tools/test/commands/analyze_once_test.dart
@@ -32,7 +32,11 @@ void main() {
     });
 
     tearDownAll(() {
-      tempDir?.deleteSync(recursive: true);
+      try {
+        tempDir?.deleteSync(recursive: true);
+      } catch (e) {
+        // ignore errors deleting the temporary directory
+      }
     });
 
     // Create a project to be analyzed