Re-write `flutter analyze` (the single-shot mode and `--flutter-repo`) to use the analysis server.

not yet ready for review -