@devoncarew 2018-04-10
OK, this is now ready for review. Previously, we had:

- `flutter analyze` implemented using the dartanalyzer command-line tool
- `flutter analyze --flutter-repo` implemented using the `package:analyzer` APIs
- `flutter analyze --watch` implemented using the analysis server

This was a lot of different code paths and implementations for doing flutter analysis. Additionally, the specific package:analyzer APIs that flutter analyze was using are deprecated, and won't support the full fidelity of the changes for the Dart 2.0 type system; it's important to move flutter analyze off of them.

This PR then:
- re-writes `flutter analyze` and `flutter analyze --flutter-repo` to be based on the analysis server
- for `flutter analyze --flutter-repo`, changes from passing 29 different directories into the analysis server to 3 top level ones; we see a substantial performance benefit from this (initial analysis of the repo using the analysis server drops from 90 seconds to 40)
- some clean up to the 3 different analysis options files in this repo to reduce duplication. We now have one top level `analysis_options.yaml` file for the repo (`analysis_options_repo.yaml` is deleted); one small delta to that in /packages/ to turn on `public_member_api_docs`, and one in /packages/flutter_tools to reference the top level options file, and not enable `public_member_api_docs`
- removed the feature of `flutter analyze`ing individual files (just directories are supported now)

Locally, I see a perf speedup in the initial (cold) repo analysis of 25% (50s to 40s), and additional (warm) analysis speedup of 3x (50s to 16s).

Since we're using the analysis server, flutter analyze is analyzing the user's pubspec.yaml file, and will now catch issues in the `flutter:` section (we have some lints for the assets section).